+++ packages/polaris-core/test/integration/server/utils/logger.ts
@@ -1,8 +1,8 @@
 import { PolarisGraphQLLogger } from '@enigmatis/polaris-graphql-logger';
-import { LoggerConfiguration } from '../../../../src';
+import { LoggerConfiguration, LoggerLevel } from '@enigmatis/polaris-logs';
 
 export const loggerConfig: LoggerConfiguration = {
-    loggerLevel: 'debug',
+    loggerLevel: LoggerLevel.DEBUG,
     writeToConsole: true,
     writeFullMessageToConsole: false,
 };
@@ -17,8 +17,9 @@ const applicationLogProperties = {
 
 export const polarisGraphQLLogger = new PolarisGraphQLLogger(
     {
-        loggerLevel: 'debug',
+        loggerLevel: LoggerLevel.INFO,
         writeToConsole: true,
+        writeFullMessageToConsole: false,
     },
     applicationLogProperties,
 );