+++ packages/polaris-graphql-logger/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+    clearMocks: true,
+    moduleFileExtensions: ['ts', 'js'],
+    preset: 'ts-jest',
+    testEnvironment: 'node',
+    testMatch: ['**/*.test.ts'],
+    coverageDirectory: './coverage/',
+};