+++ packages/polaris-typeorm/src/index.ts
@@ -1,5 +1,7 @@
-import { getFromContainer } from 'typeorm';
-import {getPolarisConnectionManager, PolarisConnectionManager} from './typeorm-bypasses/polaris-connection-manager';
+import {
+    getPolarisConnectionManager,
+    PolarisConnectionManager,
+} from './typeorm-bypasses/polaris-connection-manager';
 export { createPolarisConnection } from './typeorm-bypasses/create-polaris-connection';
 export {
     getPolarisConnectionManager,